+++ plugin-modernizer-core/src/main/java/io/jenkins/tools/pluginmodernizer/core/impl/PluginModernizer.java
@@ -35,13 +35,20 @@ public class PluginModernizer {
     private CacheManager cacheManager;
 
     /**
+     * Validate the configuration
+     */
+    public void validate() {
+        mavenInvoker.validateMavenHome();
+        mavenInvoker.validateMavenVersion();
+        ghService.validate();
+    }
+
+    /**
      * Entry point to start the plugin modernization process
      */
     public void start() {
 
-        // Validate maven
-        mavenInvoker.validateMavenHome();
-        mavenInvoker.validateMavenVersion();
+        validate();
 
         // Setup
         this.ghService.connect();
