+++ plugin-modernizer-core/src/main/resources/META-INF/rewrite/recipes.yml
@@ -10,8 +10,7 @@ recipeList:
 type: specs.openrewrite.org/v1beta/recipe
 name: io.jenkins.tools.pluginmodernizer.AddPluginsBom
 displayName: Add plugins BOM
-description: |
-  Add the Jenkins BOM to the dependenciesManagement section of the pom.xml.
+description: Add the Jenkins BOM to the dependenciesManagement section of the pom.xml.
 tags: ['chore', 'dependencies']
 recipeList:
   - org.openrewrite.jenkins.AddPluginsBom
